added Map station creation admin

+++ src/main/java/entities/Location.java
@@ -48,7 +48,7 @@ public class Location {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
@@ -56,7 +56,7 @@ public class Location {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
